LambdatoFunctions/requirements.txt
semantic-kernel==1.29.0
azure-ai-projects==1.0.0b10
azure-identity==1.22.0